+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Version 1.1.0] — 2020-04-15
+
+### Added
+
+* Include a "Limit Orders" section in the WooCommerce System Status Report ([#8]).
+* Add GitHub issue templates and contributing documentation ([#5], [#6]).
+
 ## [Version 1.0.0] — 2020-03-27
 
 Initial plugin release.
@@ -10,3 +17,7 @@ Initial plugin release.
 
 [Unreleased]: https://github.com/nexcess/limit-orders/compare/master...develop
 [Version 1.0.0]: https://github.com/nexcess/limit-orders/releases/tag/v1.0.0
+[Version 1.1.0]: https://github.com/nexcess/limit-orders/releases/tag/v1.1.0
+[#5]: https://github.com/nexcess/limit-orders/pull/5
+[#6]: https://github.com/nexcess/limit-orders/pull/6
+[#8]: https://github.com/nexcess/limit-orders/pull/8